+++ list_of_events.h
@@ -21,7 +21,7 @@ public:
 
     bool isEmpty(); // проверяет пусты ли контейнеры
 
-    uint getSize();
+    unsigned long getSize();
 
     std::set<std::string, std::less<>>::iterator
     findSetEvent(const std::string &event); // возвращаем итератор на найденное событие в set
